requirements.txt
opencv-python
numpy
flask
pillow==6.2.1
gunicorn
requests==2.22.0
tensorflow==1.13.1
argparse
google-cloud==0.34.0
google-cloud-pubsub==1.0.0
google-cloud-storage==1.19.1
google-cloud-translate==1.6.0
google-cloud-vision==0.39.0
torch==1.3.1
torchvision==0.4.2


